requirements.txt
aniso8601==7.0.0
click==8.0.1
dnspython==2.1.0
Flask==2.0.1
Flask-GraphQL==2.0.1
Flask-PyMongo==2.3.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
promise==2.3
pymongo==3.11.4
python-dotenv==0.17.1
Rx==1.6.1
six==1.16.0
Werkzeug==2.0.1